Add an option to the sindico's panel to manage new condominiums under his administration (this platform can support up to 10 condominiums). Each condominium should be registered with its relevant data.
Add an option to the syndic's panel to add council members (research online who typically forms part of the condominium council—e.g., secretary, treasurer, etc.).

In the syndic's panel:

The syndic should also have view access to the residents’ panel, but must not be allowed to change passwords—only reset them to a default one, which the user must later update to a personal one.

The syndic must have the option to register all employees of the condominium. (This app is proprietary to him and his company—it is not intended for use by other syndics.)